Fixed exporting signal when all channels are selected.
This change is connected with the channel signal selection that
can span through all channels (its channel number is -1, and
its length is arbitrary - unlike block or page signal selections).
This type of selection was not handled correctly by the export
signal functionality and it also affected the MP5 decomposition.

+++ svarog/src/main/java/org/signalml/app/view/element/SignalSelectionPanel.java
@@ -437,8 +437,8 @@ public class SignalSelectionPanel extends JPanel {
 
 		} else if (signalSelectionTypePanel.getChannelRadio().isSelected()) {
 
-			float startTime = (Float) channelSignalSelectionPanel.getStartTimeSpinner().getValue();
-			float length = (Float) channelSignalSelectionPanel.getLengthSpinner().getValue();
+			double startTime = (Double) channelSignalSelectionPanel.getStartTimeSpinner().getValue();
+			double length = (Double) channelSignalSelectionPanel.getLengthSpinner().getValue();
 
 			selection = new SignalSelection(SignalSelectionType.CHANNEL);
 			selection.setPosition(startTime);
@@ -518,6 +518,7 @@ public class SignalSelectionPanel extends JPanel {
 	 * @param errors the object in which errors are stored
 	 */
 	public void validatePanel(Errors errors) {
+
 		// there is no validation - the dialog elements enforce valid values themselves
 	}
 }
